requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
psycopg2-binary>=2.9.9
django-cors-headers>=4.3.1
python-dotenv>=1.0.0
drf-yasg>=1.21.7
pytest>=7.4.3
pytest-django>=4.7.0
factory-boy>=3.3.0
Pillow>=10.1.0
django-filter>=23.5 